base32: Moved clap argument parsing to base32.rs
Now, all base_common.rs has is the handle_input function.
